#15 ensure resources instead of just init
