+++ extension/cypress/setUp/payment_resources.js
@@ -1,7 +1,7 @@
 // Prepare all CTP resources that are needed to execute Adyen payments
 // like custom types, API integration
 
-const { initResources } = require('../../src/config/init/init-resources')
+const { ensureResources } = require('../../src/config/init/ensure-resources')
 const testUtils = require('../../test/test-utils')
 const ctpClientBuilder = require('../../src/ctp/ctp-client')
 
@@ -11,7 +11,7 @@ async function init () {
   await Promise.all([
     testUtils.deleteAllResources(ctpClient, 'types'), testUtils.deleteAllResources(ctpClient, 'extensions')
   ])
-  await initResources()
+  await ensureResources()
 }
 
 module.exports = { init }
