+++ extension/package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "scripts": {
     "test": "mocha --exit --timeout 30000 --full-trace test/**/*.spec.js",
-    "extension": "node ./extensions/src/init.js",
+    "start": "node ./src/index.js",
     "lint": "eslint .",
     "cypress:run": "cypress run",
     "cypress:open": "cypress open",
