+++ extension/resources/api-extension.json
@@ -12,5 +12,6 @@
         "Update"
       ]
     }
-  ]
+  ],
+  "timeoutInMs": 10000
 }
