+++ extension/src/config/init/ensure-api-extensions.js
@@ -2,19 +2,17 @@ const _ = require('lodash')
 
 const apiExtensionTemplate = require('../../../resources/api-extension.json')
 
-async function initApiExtensions (ctpClient, ctpAdyenIntegrationBaseUrl) {
+async function ensureApiExtensions (ctpClient, ctpAdyenIntegrationBaseUrl) {
   try {
     const extensionDraft = _.template(JSON.stringify(apiExtensionTemplate))({ ctpAdyenIntegrationBaseUrl })
-    const { body: extension } = await ctpClient.create(ctpClient.builder.extensions, extensionDraft)
-    await ctpClient.update(ctpClient.builder.extensions, extension.id, extension.version, [{
-      action: 'setTimeoutInMs',
-      timeoutInMs: 10000
-    }])
+    const { body } = await ctpClient.fetch(ctpClient.builder.extensions.where(`key="${apiExtensionTemplate.key}"`))
+    if (body.results.length === 0)
+      await ctpClient.create(ctpClient.builder.extensions, extensionDraft)
   } catch (e) {
     console.error('Error when creating API extension, skipping...', JSON.stringify(e))
   }
 }
 
 module.exports = {
-  initApiExtensions
+  ensureApiExtensions
 }
