--- extension/src/config/init/init-payment-custom-type.js
@@ -1,13 +1,0 @@
-const paymentCustomType = require('../../../resources/payment-custom-types.json')
-
-async function initPaymentCustomType (ctpClient) {
-  try {
-    await ctpClient.create(ctpClient.builder.types, paymentCustomType)
-  } catch (e) {
-    console.error('Error when creating payment custom type, skipping...', JSON.stringify(e))
-  }
-}
-
-module.exports = {
-  initPaymentCustomType
-}
