+++ extension/src/index.js
@@ -1,11 +1,11 @@
 const init = require('./server.js').setupServer()
 require('./config/config')
 
-const { initResources } = require('./config/init/init-resources')
+const { ensureResources } = require('./config/init/ensure-resources')
 
 const port = parseInt(process.env.EXTENSION_PORT || 8080, 10)
 
 init.listen(port, async () => {
-  await initResources()
+  await ensureResources()
   console.log(`Server running at http://127.0.0.1:${port}/`)
 })
